refactor: section 관련 dto 패키징 변경
