+++ src/main/java/wooteco/subway/line/controller/LineController.java
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import wooteco.subway.line.controller.dto.LineCreateDto;
-import wooteco.subway.line.controller.dto.LineDto;
+import wooteco.subway.line.service.dto.LineCreateDto;
+import wooteco.subway.line.service.dto.LineDto;
 import wooteco.subway.line.controller.dto.LineRequest;
 import wooteco.subway.line.controller.dto.LineResponse;
 import wooteco.subway.line.domain.Line;
