+++ src/main/java/wooteco/subway/line/controller/dto/LineRequest.java
@@ -4,6 +4,7 @@ import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Positive;
 import javax.validation.constraints.Size;
+import wooteco.subway.line.service.dto.LineCreateDto;
 
 public class LineRequest {
