+++ src/main/java/wooteco/subway/line/controller/dto/LineResponse.java
@@ -1,6 +1,7 @@
 package wooteco.subway.line.controller.dto;
 
 import java.util.List;
+import wooteco.subway.line.service.dto.LineDto;
 import wooteco.subway.station.controller.dto.StationResponse;
 
 public class LineResponse {
