+++ src/main/java/wooteco/subway/line/service/LineService.java
@@ -6,8 +6,8 @@ import java.util.stream.Collectors;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import wooteco.subway.exception.RequestException;
-import wooteco.subway.line.controller.dto.LineCreateDto;
-import wooteco.subway.line.controller.dto.LineDto;
+import wooteco.subway.line.service.dto.LineCreateDto;
+import wooteco.subway.line.service.dto.LineDto;
 import wooteco.subway.line.domain.Line;
 import wooteco.subway.line.domain.LineRepository;
 import wooteco.subway.section.service.dto.SectionCreateDto;
